+++ selfdrive/ui/translations/main_pl.ts
@@ -369,9 +369,9 @@
     </message>
     <message>
         <source>Get turn-by-turn directions displayed and more with a comma
-prime subscription. Sign up now: https://connect.comma.ai</source>
+prime subscription. Sign up now: https://stable.konik.ai</source>
         <translation>Odblokuj nawigację zakręt po zakęcie i wiele więcej subskrybując
-comma prime. Zarejestruj się teraz: https://connect.comma.ai</translation>
+comma prime. Zarejestruj się teraz: https://stable.konik.ai</translation>
     </message>
     <message>
         <source>No home
@@ -453,16 +453,16 @@ nie zostało ustawione</translation>
         <translation>Sparuj swoje urzadzenie ze swoim kontem comma</translation>
     </message>
     <message>
-        <source>Go to https://connect.comma.ai on your phone</source>
-        <translation>Wejdź na stronę https://connect.comma.ai na swoim telefonie</translation>
+        <source>Go to https://stable.konik.ai on your phone</source>
+        <translation>Wejdź na stronę https://stable.konik.ai na swoim telefonie</translation>
     </message>
     <message>
         <source>Click &quot;add new device&quot; and scan the QR code on the right</source>
         <translation>Kliknij &quot;add new device&quot; i zeskanuj kod QR znajdujący się po prawej stronie</translation>
     </message>
     <message>
-        <source>Bookmark connect.comma.ai to your home screen to use it like an app</source>
-        <translation>Dodaj connect.comma.ai do zakładek na swoim ekranie początkowym, aby korzystać z niej jak z aplikacji</translation>
+        <source>Bookmark stable.konik.ai to your home screen to use it like an app</source>
+        <translation>Dodaj stable.konik.ai do zakładek na swoim ekranie początkowym, aby korzystać z niej jak z aplikacji</translation>
     </message>
 </context>
 <context>
@@ -472,8 +472,8 @@ nie zostało ustawione</translation>
         <translation>Uaktualnij teraz</translation>
     </message>
     <message>
-        <source>Become a comma prime member at connect.comma.ai</source>
-        <translation>Zostań członkiem comma prime na connect.comma.ai</translation>
+        <source>Become a comma prime member at stable.konik.ai</source>
+        <translation>Zostań członkiem comma prime na stable.konik.ai</translation>
     </message>
     <message>
         <source>PRIME FEATURES:</source>
@@ -503,8 +503,8 @@ nie zostało ustawione</translation>
         <translation>comma prime</translation>
     </message>
     <message>
-        <source>CONNECT.COMMA.AI</source>
-        <translation>CONNECT.COMMA.AI</translation>
+        <source>stable.konik.ai</source>
+        <translation>stable.konik.ai</translation>
     </message>
     <message>
         <source>COMMA POINTS</source>
@@ -717,8 +717,8 @@ nie zostało ustawione</translation>
         <translation>Zakończ konfigurację</translation>
     </message>
     <message>
-        <source>Pair your device with comma connect (connect.comma.ai) and claim your comma prime offer.</source>
-        <translation>Sparuj swoje urządzenie z comma connect (connect.comma.ai) i wybierz swoją ofertę comma prime.</translation>
+        <source>Pair your device with comma connect (stable.konik.ai) and claim your comma prime offer.</source>
+        <translation>Sparuj swoje urządzenie z comma connect (stable.konik.ai) i wybierz swoją ofertę comma prime.</translation>
     </message>
     <message>
         <source>Pair device</source>